refactor: rename other files related to clerk signup
